+++ src/handlers/handleSelector/index.js
@@ -90,9 +90,19 @@ function countSelectors(selectors, report, options) {
 		report.selectors.specificity.graphData.push(tmpSpecificity);
 	}
 
-	if (compareSpecificities(report.selectors.specificity.highest, tmpSpecificity) === 1) {
-		report.selectors.specificity.highest = tmpSpecificity;
-		report.selectors.specificity.highestSelector = String(selectors);
+	report.selectors.specificity.highest10.push({
+		selector: String(selectors),
+		specificity: tmpSpecificity,
+	});
+
+	report.selectors.specificity.highest10 = report
+		.selectors
+		.specificity
+		.highest10
+		.sort((a, b) => compareSpecificities(a.specificity, b.specificity));
+
+	if (report.selectors.specificity.highest10.length > 10) {
+		report.selectors.specificity.highest10.pop();
 	}
 
 	if (baseSelectorsWithoutCombinators > options.selectorComplexityThreshold) {